Swap the order of the final two gradient FPs
Today, interpolation color space doesn't work with texture colorizers.
For that to work correctly, the colorizer will produce colors in the
final color space. (If the texture held interpolated colors, they could
fall outside the normalized range, depending on the color space).

With the texture colorizer producing values in dst, we need to skip the
interpolated to dst conversion (but not for other colorizer types).
However, we still need to convert the clamped border colors. The easiest
thing is to move the interpolated-to-dst conversion immediately after
the colorizer. This makes the colorizer step atomic: either a texture
colorizer, or a uniform colorizer + conversion step.

Then, the final (clamp, etc) FP will wrap that bundle. In that case, it
just needs to do coordinate math on the colorizer. However, the clamp
colorizer injects colors when t is outside the domain, so those colors
need to be pre-converted to the output color space. This CL extracts
some code that was already doing that work (via RP) into a helper
function, and uses that to do the conversion.

Bug: skia:13108
Bug: chromium:1449408
Change-Id: I7fb46f4513c64d3f648e61106a97fadb1b4ae353
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/709018
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>