Remove Ganesh code from SkCanvasPriv
This also moves a few Android+Ganesh specific methods to a new header
so they can be exposed only for Android.

This enforces IWYU on the affected files (ganesh/Device was
impacted because originally I was going to put the functions
in there, but decided a new file was better).

Change-Id: I96e3e557d5ccf9698fd2b747c0709f6a1444ce22
Bug: skia:14317
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/709957
Owners-Override: Kevin Lubick <kjlubick@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>