Improve WGSL codegen for break-if statements.
Previously, we would always emit `break if !(test-expression);`
with a leading logical-not hard-coded in the output. This would
generally lead to code like `break if !(i <= 4);`

Now, we will apply the logical-not to the IR, simplifying the
expression where possible. This allows us to emit the simpler
form `break if i > 4;` instead.

Bug: skia:14369
Change-Id: Ibc2b804ad96e2c3417a749940e94a1f9ef11b927
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/709019
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>