Move DDL and SkSurfaceCharacterization to include/private/chromium
This also renames them to be more clearly Ganesh-specific:
 - SkDeferredDisplayList -> GrDeferredDisplayList
 - SkDeferredDisplayListRecorder -> GrDeferredDisplayListRecorder
 - SkSurfaceCharacterization -> GrSurfaceCharacterization

SkSurface::draw(sk_sp<GrDeferredDisplayList>...) has been moved
to a static function - skgpu::ganesh::DrawDDL. This was better than
re-writing the API to take a GrDeferredDisplayList* which would
require re-wrapping in a sk_sp to be passed into the deeper layers -
an error-prone and efficient strategy.

Shim headers are still around, as well as a shim over the old
SkSurface method until Chromium can be migrated.

This CL has no functional changes in how DDLs work, besides
removing the ability to specify an offset, which was not
implemented and there are no plans to do so.

Change-Id: Ibb263d6b90e3b3616ffd11c7a03b69e78dd02fe4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/706017
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Owners-Override: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>