[Fontations] No .0.as_ref() boilerplate, simpler tag conversion
Define a with_font() method to run a closer on a FontRef returning an
Option<T> to run operations on a font or return early. This also makes
it easier to use the ?-operator and chain success/failure. Original
suggestion by Chad Brokaw.

Use built-in methods for converting from u32 to tag.

Bug: skia:14259
Change-Id: If993fedbef30c1ad062d0b11f55bb973a64a56fe
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/710636
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Dominik Röttsches <drott@google.com>