Add WGSL support for switches that do not have fallthrough cases.
The WGSL code generator now supports switch statements as long as
no cases fall through. We mirror WGSL's native abilities and
emit a comma-separated list of switch-case values when multiple
`case X:` statements are placed back-to-back. We currently tack on
a comment indicating the cases that were not handled due to
fallthrough.

In a followup CL, the non-native cases will be emulated via
if-else statements. This will mimic the rewrite-switches mode
in the GLSL code generator, but we will only rewrite the cases
that actually violate the WGSL switch rules (either by falling
through, or by being fallen-into).

Bug: skia:13092
Change-Id: I9c61399a9573a2bc57510fd4bee994f4312894f9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/710257
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>