[skif] Update morphology filters to use FilterResult
This CL completely overhauls the morphology implementation to use
FilterResult and have tighter bounds tracking.

The two main differences in bounds tracking is that the old implementation did not factor in the fact that "erode" shrinks the output bounds, and that the kernel sampling would clamp itself to the bounds of the input image. Clamping leads to inconsistencies depending on how the source image was produced (i.e. if the drawn image lands exactly on a full coverage pixel at the edge, we clamp, otherwise we introduce transparency at the boundary that was baked into the source image).

The clamping also was incorrectly applied in the old 2-pass GPU algorithm. It split the erode/dilate into an X pass and a Y pass. The initial input image would be outset by the kernel radius along both axes, but the output of the X pass was the target dimensions so all extra information for the Y pass was lost. The new shaders always use decal sampling at the source image, and preserve the extra kernel-sampled pixels across all passes.

Another major difference is how the morphology algorithms are applied. In the raster backend, there were 3 implementations (x64, neon, and regular for loops), and the GPU used two 1D passes. However, in its 1D passes it hardcoded the number of texture samples it would take. Besides that it also had specific shaders for bounds checking or no bounds checking, and erode vs. dilate.

In the new morphology filter, runtime sksl is used to consolidate CPU vs. GPU. It's implemented as two separable filters but each 1D filter may have multiple passes. There's a dense pass that applies up to a fixed, small R <= 14 using a manually unrolled switch with fallthroughs. If the radius is larger than that, the larger kernels can be constructed by sampling pixels that are farther from the current fragment, being able to double the kernel size with just 2 texture samples. There is a tradeoff between target switches, memory overhead, and reducing the number of textures. R <= 14 with the new shader pass structure is faster than the old GPU implementation on my M1 for both small and large kernels.

However, the SkSL approach is about 3x slower than the SIMD optimized for loops (and possibly the non-SIMD loop too). I'm hopeful that this reduction is acceptable, but if not, this might be the time to add direct CPU pixel-processing functions to FilterResult::Builder when the colors are 8888.

-----

This CL also fixes up some of the GMs. imagefilterscropexpand had been abusing the user bounds argument to saveLayer to artificially restrict the source image. With the old code and clamping, it wasn't obvious that this was happening, but when factoring in transparency outside the unexpectedly cropped source, it was. Using a null bounds correctly means that the image filter can erode the image that's intended to be cropped to (10,10,54,54). In this case, we see a slight change on erode in the middle row that corresponds to exactly that crop rect (which previously was hitting the case where there was no transparency in the source image so was clamping).

imagefilterscropped had nested erodes that showed the clamping effect pretty drastically (and didn't look good IMO). With just the new implementation, results were better and didn't have obvious clamping, but the nested cases looked slightly different from the (8,8) erode. This was because they had been constructed with crops that limited the requested input for the outer erode; removing those inner crops means they all match as was the intent.

The chrome layout tests have a few radical changes: https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/1427270/overview. Some show the improved edge behavior and are
easy to understand (svg/W3C-SVG-1.1/filters-morph-01-f.svg). Others are just improved invalidation bounds reporting and correctly do match what is actually changing. The one that is hardest to understand is virtual/scalefactor200/css3/filters/effect-reference-subregion-hw.html.

In the subregion test, there is an input image with a primitive
subregion on the dilate effect that excludes the colored cells.
However, the SVG spec declares that the primitive subregion is an
output crop and not an input crop, so those excluded cells should
still be factored into the dilate kernel. In the top right corner,
the excluded cells are red, so pixels should be aggregating the red
from above, the green it's on, and the blue below to turn into white.
When the old image cropped to the primitive subregion, it only
collected blue and green and produced cyan instead.

I've run this new behavior by @pdr, @fserb, and fs@opera, and their
opinion is that the expected images had just been set to what the
image filter had been producing because it was non-intuitive as to
what you should be expecting. The new results match what Firefox
produces for that test. It also highlights limitations of the SVG
spec for the morphology filters (which don't concretely define the
edge mode you need to apply). They agreed that the new behavior is
more correct.

Bug: skia:9282, chromium:400653
Bug: b/263137049
Change-Id: I3490d0968b2eddbee1cad7b5c211fe887780087a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/703457
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>