[graphite] Support ClipStack evaluation without side effects
ClipStack::applyClipToDraw() determines whether a draw is clipped out
(and can be discarded) while simultaneously updating the clip elements
with the right draw order used by clip draws during a flush. This
requires that a flush to recording required by a particular draw
operation must be performed before calling `applyClipToDraw()` for that
draw.

This leads to a chicken-and-egg problem for certain operations that may
require a flush but should be avoided if the draw is clipped out. An
example of this is an atlas draw:

   1. Atlas space for a path coverage mask should be spared for a
      clipped out shape. This means the clip stack must be queried
      before an atlas insertion.

   2. A flush to recording submits a task to render the coverage mask
      atlas and clears the tracked atlas slots for new atlas draws. The
      atlas texture render and an atlas shape draw for the same shape
      must get flushed together. This means that an atlas entry should
      be recorded after a flush.

   3. If the atlas is full, then a flush is required to clear up space
      in the atlas.

The resolve these dependencies, `applyClipToDraw()` has been split into
the `visitClipStackForDraw()` and `updateClipStateForDraw()` methods.
The former performs coarse and detailed clipping analysis by collecting
the clip elements that intersect with the draw in a complex way.
`updateClipStateForDraw()` updates these pre-computed elements and
returns the draw order that the draw should use.

Bug: b/280927575
Bug: b/246960780

Change-Id: I9f7e6802b15f503cbcd1d17aa93ae59cd8d40653
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/712637
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>