[skif] Fix morphology sksl to avoid switch fallthroughs
Fallthroughs are not supported in wgsl, so the new morphology shader
wasn't compiling on graphite+dawn. To remain ES2 compatible, this
switches to a for loop with a fixed limit and then checks the uniform
radius with an if and then breaks out.

Change-Id: I5865d7637d53bb4927d698654fdd4d79137a54c0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/712257
Commit-Queue: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>