Extract SlugImpl.h and remove #ifdefs of slugs from Canvas
I want to to remove #if defined(GANESH) etc and relocate Ganesh-
and Graphite-specific code from src/text/gpu -> src/gpu/ganesh|graphite
which will make it easier to have ganesh and graphite code both
co-exist and be configurable via a modular Bazel build.

Justification for removing #ifdefs on SkCanvas: The feature has
baked in long enough and is a no-op with the CPU backend so
there is no reason to guard it.

One day, we could have a CPU version of slugs and slugs would
be able to be moved to the sktext namespace, but today is not
that day.

Next step: Try to unify the two calls to SubRunContainer::draw
such that it doesn't take GPU specific fields, but a callback
or two to allow atlas subrun to make vertices etc.

Change-Id: I9a5663d71c02b2c43e8e38d8362d2a8c232bf017
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/711196
Reviewed-by: Herb Derby <herb@google.com>