Attempt to use fallback color type and backend format
Wtih this CL, we attempt to use the fallback color type and backend
format in SurfaceContext::asyncRescaleAndReadPixels before bailing.

credit: thanks to vasilyt@ for pointing out the color fallback code
in GrCaps and explaining that this was the reason the async rescale
and read actually succeeded and to dtapuska@ for suggested looking
at the validity of the backend format in the needsRescale block.

Bug: skia:14389
Change-Id: I81f2e18fb5cafb4466920450795ccbeddceb9808
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/712766
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Ian Vollick <vollick@chromium.org>