Disable struct-passing test on Tegra3.
This appears to be a bug in the GLSL compiler; it disallows passing
a `const T` to a function that takes `T`. In GLSL, this should be
valid because we are passing the T by value.

https://screenshot.googleplex.com/C8EEASDnLMtSNkH

Change-Id: I9988f8bc02c7c865f8af4730926ab81100303140
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/713979
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>