Graphite: Loosen TextureUsage validation to allow superset of usages
This change loosens TextureUsage validation as the BackendTexture
usages may be a superset of the TextureInfo usage within TextureProxy
when fulfilling a promise image. This is due to the read/write nature
of operations where the BackendTexture may be created with multiple
usages with write intent whereas during fulfillment the TextureProxy
and hence TextureInfo may be passed with read intent with lesser usages.

Bug: b/286044405
Change-Id: I2e12894d2ed0046890c40b665392592282cdb307
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/713736
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>