Add text and clusters to debugger text blobs
If a user provides the original text and cluster information on a text
blob then emit it in the debugger json dump. This allows the user to
know if this information is present and what it is.

Change the text blob bounds label from "coords" to "bounds".

Re-order the output of the fields. This should not make any semantic
difference but allows for easier skimming of the data in the debugger
UI.

Change-Id: I42dfa68e59260cd10f028472df8f5f8c29522966
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/714062
Commit-Queue: Herb Derby <herb@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>
Reviewed-by: Herb Derby <herb@google.com>