Add `applyRoundingHack` to `ParagraphStyle`
Using an env var for toggling the `applyRoundingHack` flag unfortunately does not work on mobile devices. Adding it to `ParagraphStyle` seem to require the least amount of plumbing and is a little bit more secretive than adding it as a Paragraph constructor argument.

I have ran the tests with `dm --match Paragraph` so the SkParagraph tests should be passing.

Change-Id: I4b366cd4a940779118919ca251aa4c9f3bfccc0f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/714308
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Weiyu Huang <weiyuhuang@google.com>