[graphite] Fix local coords in Bitmap and SDF text renderers
BitmapTextRenderStep and SDFTextRenderStep assigned `stepLocalCoords` in
their vertex shader using the potentially partially transformed atlas
bounds coordinates, which are projected relative to device space and not
the draws original local coordinates.

The vertex shaders now use the inverse local-to-device matrix to map
the device space positions of the axis-aligned quad vertices back to the
local coordinate space of the draw. The variables in the vertex shaders
have been renamed to remove the ambiguities wrt coordinate spaces.

Bug: b/288132671
Change-Id: I3903923164017006e3567cf3304e7feff6c31675
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/714817
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>