Allow use of DWriteCore
Allow the user to specify the location of dwrite headers and look in the
expected dlls for a factory.

For local use it is easiest to download the WindowsAppSDK [0] (select a
version and "Download package"). Unzip this and set skia_dwritecore_sdk
to the unzip directory. This will allow Skia to build against these
headers. For runtime, unzip the appropriate
tools/MSIX/win10-*/Microsoft.WindowsAppRuntime.*.msix and put this on
the PATH (or otherwise get DWriteCore.dll on the search path).

[0] https://www.nuget.org/packages/Microsoft.WindowsAppSDK

Change-Id: I83d2b18237a23c7f1016112cc2ac6b75eb52cf79
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/714502
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>