Move VertexFiller::fillInstanceData to src/graphite
All the implementations in AtlasSubrun subclasses just forwarded
the request to vertex filler. Thus, we can make the two callsites
bypass that intermediate call, exposing a few methods to make
that possible.

Since only graphite code is calling the graphite fillInstanceData
method, we do not need to #ifdef the implementation (or declaration)
to accomodate Ganesh-only builds.

Change-Id: I5c4e49ee08c6df3c5b92c88162897cd2b826df94
Bug: skia:14317
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/712858
Reviewed-by: Herb Derby <herb@google.com>