[skif] Update lighting image filters to use FilterResult
This is a pretty thorough rewrite. The previous implementation had
dedicated functions for performing the CPU evaluation, which have been
discarded. The GPU code used to use 9 draws (interior, 4 edges, and 4
corners) with 9 different pipelines to handle the boundary conditions.
Additionally, it specialized its pipelines for the combination of
specular vs. diffuse and distant vs. point vs. spot. There were also
subclasses of image filter and of "lights" for each type.

In the new version, the CPU implementation relies on SkSL as well.
There is only one pipeline, which has uniform branches for material
and light type (this consolidation also applies to the image filter
hierarchy and the data it stores). The edge boundary conditions are
simplifed to just clamp coordinates to a rectangle, which is much
simpler to evaluate and eventually can hopefully be mapped to HW or
a dedicated RP stage.

The SkSL for the old GPU pipelines had been spread across the various
effects and light subclasses. Now it's split into two shader
factories: one for calculating a normal map and one for evaluating the
lighting equations. My long term goal will be to have an
SkNormalMapImageFilter whose output can be automatically cached and
reused by multiple lighting image filters that share the same input
but have different parameters (and are then merged together). Having
the effects as runtime sksl makes the logic much more readable than
before, and will make such a split into two filters pretty trivial.

Change-Id: I29e6d6da026cc60bfd9e220fa0b9aa8a97c3b0bf
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/710362
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>