[fix-swo-violations] Fix sort comparison function
The compare function used to sort fNewElements assumes that its
arguments do not alias the same object. This assumption is used to
check that fNewElements does not contain duplicate elements.
However, this assumption is not guaranteed to hold by std::sort, and
the SkASSERT fails if _LIBCPP_DEBUG_STRICT_WEAK_ORDERING_CHECK is
defined.
This CL fixes gets rid of the assumption and only checks for uniqueness
if left-hand and right-hand sides of the compare function are different
objects.

Bug: b/288567882
Change-Id: Ib4430abccc922772542021dabdf15f327b3be8e0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/715857
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>