[graphite] AtlasShapeRenderStep support for inverse fills
AtlasShapeRenderStep now takes the inverseness of the fill type into
account when calculating coverage. Atlas masks are always drawn with a
regular fill and an "isInverted" flag is used to flip the coverage value
in the fragment shader.

The renderer now uses the Clip's drawBounds to calculate vertex
coordinates. The corners of the device space quad get assigned atlas
coordinates that are correctly offset relative to the actual UV
bounds of the corresponding mask. The fragment shader detects when
the interpolated UV coordinate falls outside the mask bounds (which is
common for an inverse fill) and adjusts coverage accordingly.

Bug: b/283876923
Change-Id: I330dc5ba8ea541c742c7dd03d36837f3be1811b9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/715998
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>