Reorganize crop_imagefilter GM to be easier to interpret
For testing crops with tiling, there are 4 tiling modes on the input,
4 on the output, 4 geometric relations of the input vs. the crop, and
4 geometric relations of the output to the crop.

The original GM was set up to arrange the output relations in 2x2
grids, and organize them in a column by input relation. Then there
would be a column per output tile mode, and a dedicated GM slide for
each input mode.

This was much too dense, so this reorganizes the test cases into 16
GM slides (by input tile mode and output tile mode), and then displays
the input and output geometric relations in a 4x4 grid.

Additionally, this makes tweaks to how the background image is drawn
so that it's easier to see what the blurred content should be. It also
keeps the input/crop rects intersected with the tile size for easier
triaging. When they would go offscreen, their effective images would
be filled with transparent black (correctly, since that was the extent
of the input image being drawn). But this led to unintuitive final
output that made it hard to differentiate the tile modes from each other.

Bug: skia:9296
Bug: b/263131619
Change-Id: I17662db7b444590949ba21484f52592a6fab2277
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/717176
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>