[graphite][vello] Better quality hairline strokes
Special-case hairline strokes and strokes with sub-pixel width as vello
draws these with aliasing and the results are barely visible. Draw such
strokes with a device-space width of 1 pixel and scale down the alpha by
the true width to approximate the sampled area. This produces results
that are consistent with Ganesh.

Bug: b/285351281
Change-Id: I94514f54e0920304f50dae357d3d973df426c9aa
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/718856
Commit-Queue: Arman Uguray <armansito@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>