Remove shared compiler from ResourceProvider.
This was originally added because instantiating an SkSLCompiler
was expensive; it would load and dehydrate the module data at
construction time.

Nowadays, this is no longer a concern; the SkSL Module Loader
(http://go/modules-in-sksl) ensures that modules are cached after
first use, and instantiating a Compiler is close to free. When a
Compiler is needed, you can just put one on the stack.

Change-Id: I3cc2e3720f21177f828153d99f8701e7ce874a6e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/718641
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>