[graphite] ComputePathAtlas per DrawContext
Graphite has a single ComputePathAtlas per Recorder, owned by the
Recorder's AtlasProvider. When a Device gets flushed it tells this
global atlas to record its pending masks into a DispatchGroup and clears
it, freeing it up for new mask insertions.

When a Recorder tracks multiple Devices, it is possible for one Device
to flush the global atlas and clear masks that are depended on by atlas
draws that belong to another Device that hasn't been flushed yet. The
allocated Rectanizer slots need to remain valid until all AtlasDraws
that have been recorded with them get flushed. This means that an
arbitrary flush can corrupt the atlas texture for draws of other
devices.

This bug is easily reproduced by the GM_textblobrandomfont, which
creates a second surface that leads to a premature atlas flush and
results in AtlasDraws in a RenderPassTask that reference overlapping
atlas bounds.

Solution: Each DrawContext is now responsible for lazily instantiating
its own separate ComputePathAtlas. This restricts a Device to interact
only with the atlas that it records AtlasShape draws to and when a
DrawContext gets flushed it doesn't interfere with other DrawContexts'
atlases. PathAtlases still all render to the same TextureProxy while
maintaining separate Rectanizers, which avoids increasing GPU resource
use.

Change-Id: I2f289a361f996beb9404c868c95a193d77667263
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/719316
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>