[graphite] Add more support in asyncrescaleandread GMs.
Adds in Graphite code for the non-YUV asyncrescaleandread GMs.
Also fixes an issue with non-YUV code, where it was trying to
transfer the data to the CPU prior to performing the rescale.
On a Recorder snap it adds the copies before the draws, so we
have to presnap to ensure the draws come first.

Bug: b/290198076
Change-Id: Ibe3f236f86088a48937b237a490a15209e1be278
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/720061
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Jim Van Verth <jvanverth@google.com>