Pass up insert failure for SkTriangulator
Pass up a failure for inserting an edge in the active set.
Wire this into existing failure system.

It looks like we have fixed the remove problems, and now
the insert problems have popped up. Time to whack!

This looks like it fixes two of the regressions.

Bug: oss-fuzz:60466
Bug: oss-fuzz:60504
Bug: skia:14599
Change-Id: I137d22d55e8cef17901ddbe0d0b7b4d5d9fcb2d3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/720909
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Herb Derby <herb@google.com>