Add new (internal) SkCanvas drawAsTiledImageRect virtual
In order for DrawImageRect_Ganesh to operate at the SkCanvas level we need to provide a means for the drawImage/drawImageRect calls to be intercepted before devolving the draw to the Device.

Note that this change has the side effect that the Ganesh no longer does tiling for drawImageLattice and drawEdgeAAImageSet calls (since the tiling is now pulled out of Device::drawImageRect).

This CL plumbs the SkCanvas down but doesn't make use of it yet.

Bug: b/267656937
Change-Id: Ie605f4de419da7238254e9aa5940c7927c3851d2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/721897
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>