[skif] Replace SkTileImageFilter with nested crops
The visual semantics of the SkImageFilters::Tile factory are cropping
the input to the src rect, and repeat tiling it into the dst rect,
which is then decal cropped.

This can be represented with two SkCropImageFilters and is likely
more efficient with the tile mode application (as a shader when
possible, vs. always rendering images).

Bug: skia:9282
Bug: b/263137785
Change-Id: I9ad3305e5022757bd98df5ad0e17a245708a0fe8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/721198
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>