Ensure index-substitution expressions are initialized before use.
The workaround in http://review.skia.org/722477 did not account for
expression evaluation-order; it worked in practice, but there was no
guarantee that the index-substitution variable would be initialized
before use. This CL uses a sequence-expression to force the
initialization to occur in the proper order.

Bug: skia:14130
Change-Id: Iad05d063115288a2c15ddfdefd8e373317a3538d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/723337
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Arman Uguray <armansito@google.com>
Commit-Queue: Arman Uguray <armansito@google.com>